Adjusted constraint on use of qudt:unit property
Overconstraint on requiring qudt:Quantity to have a qudt:unit value. This constraint is inherited from qudt:Quantifiable. This is only true if the qudt:value property is populated. A qudt:Quantity should not have a value if the qudt:quantityValue object property is used instead.
For now, I am moving the cardinality of 1 constraint from qudt:Quantifiable to qudt:QuantityValue.
This impacts our qudt:PhysicalConstant instances.
In the future, we can define a more sophisticated SHACL constraint for qudt:Quantity that would check whether qudt:value is populated, and only then require a qudt:unit value.