Remove erroneous skos:broader triples
When adding qudt:exactMatch, a number of quantitykinds did not have their skos:broader triples removed. This commit removes them (assuming that in all these cases, the qudt:exactMatch triple was correctly added).

Also, a check for this constellation is added to the collections/COLLECTION_QUDT_QA_TESTS_ALL-v2.1.ttl test file. Moreover, the node shapes in that file is adapted to include an explicit `targetClass` triple as this makes it easier to validate with those files - no need to import the ontology.