Added in the additional units from the project
I added in the Osmotic Concentration, since it seemed to fit with the other quantities.  I only added the unit we needed not the whole base and derived is this acceptable?

We needed a couple more units combinations of existing units.

I added the units for the endotoxins as per the USP 85 that we are using, I didn't make a new quantity as it seems not needed as you have a count which is what this would be although you do have Microbial Formation which seems a little specialised?