Fix spotless config (#1250)
* Fix the spotless plugin config

It is important that no file is configured to be formatted by multiple formatters - otherwise the results are undefined, and possibly unexpected.

Now, the default configuration for RDF only covers src files and has only the turtle-formatter fix (removing `\r`) characters.

The execution `format-dist` modifies this to apply only to ttl files in dist and replace all the placeholders.

* Make sure spotless runs before shacl

* Fix src formatting

* Update CHANGELOG.md

* Improve fix workflow